switching os'es

+++ src/core/types.hpp
@@ -5,8 +5,8 @@
 #include <cstdint>
 
 struct TilePosition {
-	int16_t x;
-	int16_t y;
+	int32_t x;
+	int32_t y;
 
 	inline Vector2 RaylibVector() {
 		return Vector2{static_cast<float>(x), static_cast<float>(y)};
